Fix ipv6 support
There is issue with ipv6 where address has to be in brackets, this causes the
underlying ruby TCPSocket to fail. Net::HTTP.new will fail without brackets on
jojning the ipv6 address with :port or passing brackets to TCPSocket. It was
found that if we use Net::HTTP.start with url.hostname the incriminated code
won't be hit.

Change-Id: I99bbb8b7aaeae8958f757772ba6186f1e48a6778
Closes-Bug: rhbz#1185652